+++ test/azure/Expected/AcceptanceTests/Paging/paging/operations/_paging_operations.py
@@ -1071,6 +1071,9 @@ class PagingOperations(object):
             **kwargs
         )
 
+        kwargs.pop('error_map', None)
+        kwargs.pop('content_type', None)
+
         def get_long_running_output(pipeline_response):
             deserialized = self._deserialize('ProductResult', pipeline_response)